backend/requirements.txt
fastapi
uvicorn
pillow
opencv-python-headless
numpy
torch
timm
facenet-pytorch
python-multipart
